Change the way Docker Images are built
When using the installer, each time the Images are built, a new Image with new ID is created, ever if the installer and the base Images are the same.

This causes new Images to be created whenever I push a commit to this repo, although the change might be unrelated to some of the Images.

By taking advantage of Docker multi-stage builds and ASP.NET runtime binary files, each time the Images are built, they produce Images with the same ID, if their `dockerfile` was not changed.